fix black/friends

+++ DatingWebsiteApp/App.BLL/Services/BlackListService.cs
@@ -105,8 +105,11 @@ namespace App.BLL.Services
                     UserFromId = user_id,
                     UserToId = bad_guy_id
                 });
-
-                await _friendService.DeleteFriendAsync(user_id, bad_guy_id);
+                var friendship = _friendService.GetFriendshipFor(user_id, bad_guy_id);
+                if (friendship != null)
+                {
+                    await _friendService.DeleteFriendAsync(user_id, bad_guy_id);
+                }
                 var new_bad = new UserTabVM(bad_guy);
                 return new_bad;
             }
